+++ src/bits/boolean/mod.rs
@@ -1,6 +1,6 @@
 use ark_ff::{BitIteratorBE, Field, PrimeField};
 
-use crate::{fields::fp::FpVar, prelude::*, ToConstraintFieldGadget, Vec};
+use crate::{fields::fp::FpVar, prelude::*, Vec};
 use ark_relations::r1cs::{
     ConstraintSystemRef, LinearCombination, Namespace, SynthesisError, Variable,
 };
@@ -9,6 +9,7 @@ use core::borrow::Borrow;
 mod allocated;
 mod and;
 mod cmp;
+mod convert;
 mod eq;
 mod not;
 mod or;
@@ -19,10 +20,6 @@ pub use allocated::AllocatedBool;
 
 #[cfg(test)]
 mod test_utils;
-
-pub(crate) fn bool_to_field<F: Field>(val: impl Borrow<bool>) -> F {
-    F::from(*val.borrow())
-}
 
 /// Represents a boolean value in the constraint system which is guaranteed
 /// to be either zero or one.
@@ -128,6 +125,8 @@ impl<F: Field> Boolean<F> {
 
     /// Convert a little-endian bitwise representation of a field element to
     /// `FpVar<F>`
+    ///
+    /// Wraps around if the bit representation is larger than the field modulus.
     #[tracing::instrument(target = "r1cs", skip(bits))]
     pub fn le_bits_to_fp(bits: &[Self]) -> Result<FpVar<F>, SynthesisError>
     where
@@ -178,89 +177,6 @@ impl<F: Field> Boolean<F> {
             Ok(combined)
         }
     }
-
-    /// Enforces that `bits`, when interpreted as a integer, is less than
-    /// `F::characteristic()`, That is, interpret bits as a little-endian
-    /// integer, and enforce that this integer is "in the field Z_p", where
-    /// `p = F::characteristic()` .
-    #[tracing::instrument(target = "r1cs")]
-    pub fn enforce_in_field_le(bits: &[Self]) -> Result<(), SynthesisError>
-    where
-        F: PrimeField,
-    {
-        // `bits` < F::characteristic() <==> `bits` <= F::characteristic() -1
-        let mut b = F::characteristic().to_vec();
-        assert_eq!(b[0] % 2, 1);
-        b[0] -= 1; // This works, because the LSB is one, so there's no borrows.
-        let run = Self::enforce_smaller_or_equal_than_le(bits, b)?;
-
-        // We should always end in a "run" of zeros, because
-        // the characteristic is an odd prime. So, this should
-        // be empty.
-        assert!(run.is_empty());
-
-        Ok(())
-    }
-
-    /// Enforces that `bits` is less than or equal to `element`,
-    /// when both are interpreted as (little-endian) integers.
-    #[tracing::instrument(target = "r1cs", skip(element))]
-    pub fn enforce_smaller_or_equal_than_le<'a>(
-        bits: &[Self],
-        element: impl AsRef<[u64]>,
-    ) -> Result<Vec<Self>, SynthesisError>
-    where
-        F: PrimeField,
-    {
-        let b: &[u64] = element.as_ref();
-
-        let mut bits_iter = bits.iter().rev(); // Iterate in big-endian
-
-        // Runs of ones in r
-        let mut last_run = Boolean::constant(true);
-        let mut current_run = vec![];
-
-        let mut element_num_bits = 0;
-        for _ in BitIteratorBE::without_leading_zeros(b) {
-            element_num_bits += 1;
-        }
-
-        if bits.len() > element_num_bits {
-            let mut or_result = Boolean::constant(false);
-            for should_be_zero in &bits[element_num_bits..] {
-                or_result |= should_be_zero;
-                let _ = bits_iter.next().unwrap();
-            }
-            or_result.enforce_equal(&Boolean::constant(false))?;
-        }
-
-        for (b, a) in BitIteratorBE::without_leading_zeros(b).zip(bits_iter.by_ref()) {
-            if b {
-                // This is part of a run of ones.
-                current_run.push(a.clone());
-            } else {
-                if !current_run.is_empty() {
-                    // This is the start of a run of zeros, but we need
-                    // to k-ary AND against `last_run` first.
-
-                    current_run.push(last_run.clone());
-                    last_run = Self::kary_and(&current_run)?;
-                    current_run.truncate(0);
-                }
-
-                // If `last_run` is true, `a` must be false, or it would
-                // not be in the field.
-                //
-                // If `last_run` is false, `a` can be true or false.
-                //
-                // Ergo, at least one of `last_run` and `a` must be false.
-                Self::enforce_kary_nand(&[last_run.clone(), a.clone()])?;
-            }
-        }
-        assert!(bits_iter.next().is_none());
-
-        Ok(current_run)
-    }
 }
 
 impl<F: Field> From<AllocatedBool<F>> for Boolean<F> {
@@ -283,28 +199,10 @@ impl<F: Field> AllocVar<bool, F> for Boolean<F> {
     }
 }
 
-impl<F: Field> ToBytesGadget<F> for Boolean<F> {
-    /// Outputs `1u8` if `self` is true, and `0u8` otherwise.
-    #[tracing::instrument(target = "r1cs")]
-    fn to_bytes(&self) -> Result<Vec<UInt8<F>>, SynthesisError> {
-        let value = self.value().map(u8::from).ok();
-        let mut bits = [Boolean::FALSE; 8];
-        bits[0] = self.clone();
-        Ok(vec![UInt8 { bits, value }])
-    }
-}
-
-impl<F: PrimeField> ToConstraintFieldGadget<F> for Boolean<F> {
-    #[tracing::instrument(target = "r1cs")]
-    fn to_constraint_field(&self) -> Result<Vec<FpVar<F>>, SynthesisError> {
-        let var = From::from(self.clone());
-        Ok(vec![var])
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::Boolean;
+    use crate::convert::ToBytesGadget;
     use crate::prelude::*;
     use ark_ff::{BitIteratorBE, BitIteratorLE, Field, One, PrimeField, UniformRand};
     use ark_relations::r1cs::{ConstraintSystem, SynthesisError};
@@ -398,87 +296,6 @@ mod test {
             Boolean::enforce_in_field_le(&bits)?;
 
             assert!(cs.is_satisfied().unwrap());
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_enforce_nand() -> Result<(), SynthesisError> {
-        {
-            let cs = ConstraintSystem::<Fr>::new_ref();
-
-            assert!(
-                Boolean::enforce_kary_nand(&[Boolean::new_constant(cs.clone(), false)?]).is_ok()
-            );
-            assert!(
-                Boolean::enforce_kary_nand(&[Boolean::new_constant(cs.clone(), true)?]).is_err()
-            );
-        }
-
-        for i in 1..5 {
-            // with every possible assignment for them
-            for mut b in 0..(1 << i) {
-                // with every possible negation
-                for mut n in 0..(1 << i) {
-                    let cs = ConstraintSystem::<Fr>::new_ref();
-
-                    let mut expected = true;
-
-                    let mut bits = vec![];
-                    for _ in 0..i {
-                        expected &= b & 1 == 1;
-
-                        let bit = if n & 1 == 1 {
-                            Boolean::new_witness(cs.clone(), || Ok(b & 1 == 1))?
-                        } else {
-                            !Boolean::new_witness(cs.clone(), || Ok(b & 1 == 0))?
-                        };
-                        bits.push(bit);
-
-                        b >>= 1;
-                        n >>= 1;
-                    }
-
-                    let expected = !expected;
-
-                    Boolean::enforce_kary_nand(&bits)?;
-
-                    if expected {
-                        assert!(cs.is_satisfied().unwrap());
-                    } else {
-                        assert!(!cs.is_satisfied().unwrap());
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_kary_and() -> Result<(), SynthesisError> {
-        // test different numbers of operands
-        for i in 1..15 {
-            // with every possible assignment for them
-            for mut b in 0..(1 << i) {
-                let cs = ConstraintSystem::<Fr>::new_ref();
-
-                let mut expected = true;
-
-                let mut bits = vec![];
-                for _ in 0..i {
-                    expected &= b & 1 == 1;
-                    bits.push(Boolean::new_witness(cs.clone(), || Ok(b & 1 == 1))?);
-                    b >>= 1;
-                }
-
-                let r = Boolean::kary_and(&bits)?;
-
-                assert!(cs.is_satisfied().unwrap());
-
-                if let Boolean::Var(ref r) = r {
-                    assert_eq!(r.value()?, expected);
-                }
-            }
         }
         Ok(())
     }
