+++ src/fields/nonnative/allocated_field_var.rs
@@ -3,7 +3,11 @@ use super::{
     reduce::{bigint_to_basefield, limbs_to_bigint, Reducer},
     AllocatedNonNativeFieldMulResultVar,
 };
-use crate::{fields::fp::FpVar, prelude::*, ToConstraintFieldGadget};
+use crate::{
+    convert::{ToBitsGadget, ToBytesGadget, ToConstraintFieldGadget},
+    fields::fp::FpVar,
+    prelude::*,
+};
 use ark_ff::{BigInteger, PrimeField};
 use ark_relations::{
     ns,
